Initial commit.
Add bootsector bare bones. For now, It just displays a loading
message and recalibrate the hard-disk read/write head position
using BIOS interrupts.

This bootsector can act as a single-partition hard-disk mbr or
be installed in the first sector of the OS partition where it
can be chainloaded using grub (or an equivalent).

Native multi-boot spec implementation may be added in the future;
I wanted to gain some experience in x86 real-mode assembly and
bios interrupts first.