I discovered that in most modern BIOSes, their extended data area spans to big chunks of memory. It's generally advised not to load anything in the [0x90000 - 0x9ffff] area as modern BIOSes usually occupy that region; we'll typically avoid it.
The mbr's final result is loading the kernel at the [0 - 0x80000]
region in a protected mode flat memory model environment where all
the segments spans the entire 4G virtual memory region and no
paging yet enabled.

We won't load the kernel directly from disk to 0x00 as this will
strike out the BIOS and the real-mode IVT while they're still in
need reading disk sectors. We first load the kernel to the free
[0x00500-0x804ff] region, then after being sure we won't need the
bios anymore, we copy the kernel code to the region beginning from
physical address 0x00.

On the very start, we'll also relocate the mbr to 0x80500 not to
let the mbr overwrite itself while loading the kernel from disk
sectors. This means the mbr will default to real-mode code and
data segment 0x8050; we issue a far jump to this segment directly
after relocation.