Several faults in the inline assembly code of memset.
+ First, we used all of the input operands as outputs without
telling gcc:

"There is no way for you to specify that an input operand is
 modified without also specifying it as an output operand. "

+ Second, all of our otuput operands where for the sake of
telling gcc that they're clobbered, no more. Thus, a "volatile"
was needed:

"Note that if all the output operands you specify are for this
purpose (and hence unused), you will then also need to specify
volatile for the asm construct, as described below, to prevent
GCC from deleting the asm statement as unused."

+ Third, we have several inputs and outputs operands. We do not
use the inputs, throw them, and then use the outputs in order,
thus the `&' constraint modifier was needed:

"Unless an output operand has the `&' constraint modifier, GCC
may allocate it in the same register as an unrelated input
operand, on the assumption the inputs are consumed before the
outputs are produced. This assumption may be false if the
assembler code actually consists of more than one instruction.
In such a case, use `&' for each output operand that may not
overlap an input."

That's infact what has exactly happened. We store a value in
%[ch] (input, the "ch" parameter), then use %[tmp] as output,
then use the previously stored value in %ch.

GCC assumed that cause we already used %ch, we no longer needed
it. Thus, it allocated the same register we use for %ch (%rax)
to the output %tmp. This is, ofcourse, disastrous.

+ Fourth, memset should return the given pointer unmodified. To
do so, we bind %[dst] to an unused pointer in output mode (d0),
and to the real parameter value in input mode (dst).

This leads gcc to emit the right output:

  70:   49 89 f8                mov    %rdi,%r8
  <play with %rdi here>
  9d:   4c 89 c0                mov    %r8,%rax
  a0:   c3                      retq