Introduce Intel's MP Specification support.
Get information about legacy ISA IRQs mapping to I/O APIC pins,
the number of usable I/O APICs and their base addresses, among
several other important system-wide details.

Note that while the Intel MP spec is so sweet and still in
fact supported in these days machines, MOST x86-64 systems are
designed with ACPI as the primary model. Everyone knows that
ACPI designers should've been buried alive, but we have to
live with it.

Preliminary ACPI support is to be added in a very later stage
once the need to write the PCI code arise.