Detect where the i8259 INTR pin is connected to the IOAPICs.
We use two detection methods, the first is examining the IOAPICs
redirection table entries for an unmasked entry of delivery mode
ExtINT. If this entry is found, then we found the pin.

The second method is scanning the mptables for an IOINTERRUPT
ISA entry representing the i8259 pin. It's usually pin #0 in
one of the ioapics.

In my laptop, scanning the IOAPIC entries didn't return any
results, but the MP-table had the pin position which was IOAPIC