Multiple-Processor (MP) Initialization.
Fire up non-bootstrap AP cores in the system using the global
INIT-SIPI-SIPI method. The AP cores start in real-mode as any
other x86 chip; the job of the 'trampoline' is to jump the AP
cores to long mode, and to the rest of kernel C code.

Additionally, the trampoline also sets up the GDT, IDT, and
page tables of those cores.