Introduce safe for IRQs-context spinlocks.
Regular spinlocks can lead to deadlock situations when an IRQ
handler tries to grab a lock already held by the interrupted core.
For such cases, we use spin_lock_irqsave() which disables local
interrupts before acquiring the lock.

A use case is the keyboard handler and the vga code, where the
common deadlock case is fixed.