Introduce the Kernel Memory Allocator!
I've managed to steal a good number of days between the exams to
finish this.

A lot of papers were surveyed to choose the algorithm that would
fit us best. The algorithms checked included K&R-style memory
allocators, power-of-2 lists, McKusick-Karels, Paul McKenney's
SMP-friendly allocators, Mach Zone, Solaris Slab, and Linux Slub.

The algorithm chosen is the 4.3BSD's McKusick-Karels allocator for
its elegance, simplicity, super-fast O(1) access, and good memory
utilization for small allocations. The algorithm obviously also
suffers from a number of drawbacks discussed in the comments.

We shall implement a slab-like caching allocator for big objects
once the need arise. We're obviously more influenced by FreeBSD
in this area than any other UNIX implementation.

This is one of the most interesting parts of the project so far ..