Rename x86-64's PDP and PD to PML3 and PML2 respectively.
At first there were only two levels in the page translation hierarchy:
the Page Directory and the Page Table. For PAE support in P6+ CPUs,
Intel increased the virtual space to 36 bits and added the Page
Directory Pointer. AMD added yet another hierarchy level for 64-bit
support, the PML4 table, and kept older names for the lower levels.

Of course current situation is a mess, so we'll just use 'PMLi' names
for the translation tables, where i is table's level.

This change is in preparation for our major kernel address-space
changes to be apparent in some future commits.