Bugfix: Avoid emitting code that uses the AMD64 ABI 'red zone'
Basically the zone is a 128-byte area _below_ the stack pointer
that can be used for temporary local state, especially for leaf
functions.

Of Course this red zone is disastrous to be used in the kernel
since it's where the CPU pushes %ss, %rsp, %rflags, %cs, and %rip
on the event of invoking an interrupt handler. If used, it will
simply lead to kernel stack corruption.

Check attached notes for extra context on the debugging horrors
faced to catch this bug!