Remove trickery used to save 28-KBytes of final executable size.
Those 28-KBytes were, in fact, the bootstrap BSS section size.

We used some trickery of setting the kernel Load Memory Address
(LMA) to bootstrap BSS section start load address. This way, we
avoided any allocation of final image size to bootstrap's BSS.

The downside is the need of copying the kernel image away from
bootstrap BSS to where it was originally linked to on the very
early start. This was a very risky effort, especially cause we
needed to copy the kernel image to a 'safe haven' first, to
avoid deadly overlaps while doing kernel image copying.

The trickery was quite complex; I feel good about getting rid
of it. Most kernels are distributed in gzipped format anyway,
which makes the size effects of a 28KBs of zeroes a non-issue.