Bugfix: Reload segment caches directly upon switching to long-mode.
We forgot to update the segment caches with the values of the new
64-bit GDT upon reaching long mode.

That was why we didn't catch a previous bug in the %ds segment
descriptor until we enabled interrupts (where stack and data
segmentswe are setup with the same descriptor, and the stack
segment get reloaded after each `iret').

We now nullify all the segments of no use to long mode. This
includes all x86 segments except %cs.