Strings: Assert against memcpy() overlaps.
memcpy_forward(): a new function to __explicitly__ support
copying from and to overlapping regions when src > dst. This
was needed in the VGA screen scrolling code.

While we were at it, the string ops are now C99-compliant.
The length parameter is of 'size_t' type, and the 'restrict'
keyword is now used in the prototypes as advised.

There's also a memcpy minus the checks for kernel hot paths.