Introduce Ramdisk Support!
Provide a Python script for building the final Cute disk image,
which consists of a (possibly-extended) 512KB kernel, a ramdisk
header that is built by the script, and the ramdisk itself.

For each ramdisk sector, we switch back to real-mode, load that
sector using BIOS extended-readed EDD service, then return back
to protected mode and re-copy it above the 1MB low-memory area!

The Final result is  a  ramdisk which is loaded  directly after
the kernel's binary image.  This code has been tested on 10/ten
boxes without visible problems;  some AMI and Phoenix BIOS bugs
are mentioned in the comments.