Makefile: Automatically build Cute's final disk image.
After a single make, the final disk image (which may include
a ramdisk) will  reside in 'build/hd-image'.  'hd-image' can
then be written directly to a USB drive, or treated as a ful-
ly bootable disk in an x86-64 virtual machine.

Most of the hard work was done in the python script included
in the previous commit.